requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
langchain_chroma
langgraph
python-dotenv
pydantic
fastapi
beautifulsoup4
uvicorn
langserve
wget
transformers
torch
bitsandbytes
huggingface_hub
accelerate
sentence-transformers
hf_xet
pypdf
rapidocr-onnxruntime
